requirements.txt
google-adk
requests
beautifulsoup4
tqdm
python-dotenv
vertexai